+++ lib/Controller/ImageController.php
@@ -180,11 +180,10 @@ class ImageController extends GenericApiController
         int $id,
         bool $basic = false,
         bool $current = false,
-        bool $filename = false,
         bool $tags = false,
         string $clusters = ''
     ): Http\Response {
-        return Util::guardEx(function () use ($id, $basic, $current, $filename, $tags, $clusters) {
+        return Util::guardEx(function () use ($id, $basic, $current, $tags, $clusters) {
             $file = $this->fs->getUserFile($id);
 
             // Get the image info
@@ -205,6 +204,14 @@ class ImageController extends GenericApiController
             // Allow these ony for logged in users
             $user = $this->userSession->getUser();
             if (null !== $user) {
+                { // Get the path of the file if in current user's files
+                    $path = $file->getPath();
+                    $parts = explode('/', $path);
+                    if (\count($parts) > 3 && $parts[1] === $user->getUID()) {
+                        $info['filename'] = $path;
+                    }
+                }
+
                 // Get list of tags for this file
                 if ($tags) {
                     $info['tags'] = $this->getTags($id);
@@ -213,14 +220,6 @@ class ImageController extends GenericApiController
                 // Get latest exif data if requested
                 if ($current) {
                     $info['current'] = Exif::getExifFromFile($file);
-                }
-
-                // Get the path of the file for the current user
-                if ($filename) {
-                    $parts = explode('/', $file->getPath());
-                    if (\count($parts) > 3 && $parts[1] === $user->getUID()) {
-                        $info['filename'] = '/'.implode('/', \array_slice($parts, 3));
-                    }
                 }
 
                 // Get clusters for this file