+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+- **Feature**: Support showing full file path in sidebar ([#173](https://github.com/pulsejet/memories/issues/173))
+- **Feature**: View file in folder on clicking name in sidebar
 - **Fix**: Support for transcoding MKV files.
 
 ## [v5.4.1] - 2023-08-20